requirements.txt
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-community==0.3.24
langchain-pinecone>=0.2.0
pinecone>=5.0.0
streamlit==1.45.1
beautifulsoup4==4.13.4
python-dotenv==1.1.0
unstructured==0.17.2
pypdf==5.6.0
